requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.1
bcrypt==4.0.1
pyjwt==2.8.0